backend/requirements.txt
flask==3.0.0
gunicorn==21.2.0
pandas==2.1.4
openpyxl==3.1.2
customtkinter
requests==2.31.0
python-dotenv==1.0.0